requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
Flask-Bcrypt==1.0.1
Flask-Migrate==3.1.0
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask==2.1.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
pip==21.3.1
pkgutil-resolve-name==1.3.10
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyrsistent==0.18.1
pytz==2022.1
setuptools==60.2.0
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
wheel==0.37.1
zipp==3.8.1
cryptography==37.0.4

